64-bit fixes for sipXportLib tests.
git-svn-id: http://scm.sipfoundry.org/rep/sipX/main@10946 a612230a-c5fa-0310-af8b-88eea846685b

+++ sipXportLib/src/test/os/TurnMessageTest.cpp
@@ -1,8 +1,12 @@
+// OS INT TYPES
+#include "os/OsIntTypes.h"
+
 #include "TurnMessageTest.h"
 #include "os/TurnMessage.h"
 #include <cppunit/CompilerOutputter.h>
 #include <cppunit/extensions/TestFactoryRegistry.h>
 #include <cppunit/ui/text/TestRunner.h>
+#include <string.h>
 
 /*
 int main(int argc, char* argv[])
@@ -196,7 +200,7 @@ void TurnMessageTestSuite::testTurnEncodeParseBasic()
     bool bRC ;
     char szString[1024] ;
     unsigned short usValue ;
-    unsigned long ulValue ;
+    uint32_t ulValue ;
     char* ptr ;
 
     // Valid empty state
